Process dropped items in UploadCard.
Adds a callback to UploadCard called dropFn() that gets passed the event
and a list of items that were dropped into the card.

Adds an implementation of the callback that processes the objects and
generates the state of the file listing.

The file list is now dynamic.

Re-org the theme a bit to allow for changing the border color of the
UploadCard.